+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+## Unreleased
+
+### Added
+
+- Support for users that require multi-factor authentication.
+
 ## 1.5.1 - 2021-05-12
 
 ### Fixed
@@ -20,8 +26,6 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 ## 1.5.0 - 2021-05-05
 
 ### Added
-
-- Support for users that require multi-factor authentication.
 
 - New command `code42 alerts show` that displays information about a single alert.