+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 The intended audience of this file is for py42 consumers -- as such, changes that don't affect
 how a consumer would use the library (e.g. adding unit tests, updating documentation, etc) are not captured here.
 
+
 ## Unreleased
 
 ### Added
@@ -15,6 +16,12 @@ how a consumer would use the library (e.g. adding unit tests, updating documenta
 - New command `code42 users add-role` to add a user role to a single user.
 
 - New command `code42 users remove-role` to remove a user role from a single user.
+
+## 1.6.0 - 2021-05-20
+
+### Added
+
+- Support for users that require multi-factor authentication.
 
 ## 1.5.1 - 2021-05-12